[refactor] addCrime과 addCrimes 리팩토링
- addCrime과 addCrimes의 중복 코드 함수화
   두 api 모두 하나의 crime을 추가하는 동일한 로직을 사용하므로
   service 계층에 함수를 추가하여 중복 코드 제거
