+++ src/routes/sdk/controllers/addCrimes.ts
@@ -1,8 +1,7 @@
 /* eslint-disable no-param-reassign */
 import { Context, Next } from 'koa';
-import { Types } from 'mongoose';
-import Crime, { ICrime, ICrimeDocument } from '../../../models/Crime';
-import Issue from '../../../models/Issue';
+import { ICrime } from '../../../models/Crime';
+import addCrime from '../services/addCrime';
 
 export default async (ctx: Context, next: Next): Promise<void> => {
   const newCrimes: ICrime[] = ctx.request.body;
@@ -15,24 +14,7 @@ export default async (ctx: Context, next: Next): Promise<void> => {
   try {
     await Promise.all(
       newCrimes.map(async (newCrime) => {
-        const newCrimeDoc: ICrimeDocument = Crime.build(newCrime);
-        const res = await newCrimeDoc.save();
-        await Issue.findOneAndUpdate(
-          {
-            projectId: Types.ObjectId(projectId),
-            message: newCrime.message,
-            stack: newCrime.stack,
-            type: newCrime.type,
-            isOpen: true,
-          },
-          {
-            $push: { crimeIds: res._id },
-          },
-          {
-            new: true,
-            upsert: true,
-          },
-        );
+        await addCrime(newCrime, projectId);
       }),
     );
     ctx.response.status = 200;
